requirements.txt
beautifulsoup4~=4.12.3
jsonschema
lxml
requests
click>=8.1
numpy~=1.23.0
pillow~=10.3.0
regex
scipy~=1.10.1
protobuf>=3.0.0
coremltools~=6.0
Jinja2~=3.0
python-bidi
torchvision>=0.5.0
torch~=2.0.1
scikit-learn~=1.2.1
scikit-image>=0.17.0,<=0.19.3
Shapely~=1.8.5
pyarrow
pytorch-lightning~=2.0.0
torchmetrics>=0.10.0
rich